Simplify iconItem and remove iconItemWithLink
